cup-simulator/unittests/requirements.txt
protobuf>=4.21.0
paho-mqtt>=1.6.0

